requirements.txt
streamlit
openai
python-dotenv
requests
feedparser
googlenewsdecoder